refactore the backend
