+++ client/src/utils/mutation.js
@@ -4,49 +4,13 @@ export const LOGIN_USER = gql`
   mutation login($email: String!, $password: String!) {
     login(email: $email, password: $password) {
       token
-      contractor {
+      user {
         _id
-        name
+        userName
       }
     }
   }
 `;
-export const ADD_CONTRACTOR = gql`
-  mutation addContractor(
-    $name: String!
-    $email: String!
-    $password: String!
-    $typeOfServices: String!
-    $address: String!
-    $driverLicense: String!
-    $cert: String
-    $serviceZipcode: String!
-    $bio: String
-    $photo: String
-    $phoneNumber: String!
-  ) {
-    addContractor(
-      name: $name
-      email: $email
-      password: $password
-      typeOfServices: $typeOfServices
-      address: $address
-      driverLicense: $driverLicense
-      cert: $cert
-      serviceZipcode: $serviceZipcode
-      bio: $bio
-      photo: $photo
-      phoneNumber: $phoneNumber
-    ) {
-      token
-      contractor {
-        _id
-        name
-      }
-    }
-  }
-`;
-
 export const ADD_USER = gql`
   mutation addUser(
     $userName: String!
@@ -56,6 +20,9 @@ export const ADD_USER = gql`
     $password: String!
     $phoneNumber: String!
     $address: String!
+    $role:String!
+    $typeOfServices:String
+    $serviceZipcode:String
   ) {
     addUser(
       userName: $userName
@@ -65,11 +32,16 @@ export const ADD_USER = gql`
       password: $password
       phoneNumber: $phoneNumber
       address: $address
+      role:$role
+      typeOfServices:$typeOfServices
+      serviceZipcode:$serviceZipcode
     ) {
       token
-      User {
+      user {
         _id
-        name
+        userName
+        email
+        role
       }
     }
   }
