+++ server/models/index.js
@@ -1,5 +1,4 @@
-const Contractor = require("./Contractor");
 const User = require("./User");
 const Createproject = require("./Createproject");
 
-module.exports = { Contractor, User, Createproject };
+module.exports = { User, Createproject };
