+++ server/schemas/typeDefs.js
@@ -1,17 +1,6 @@
 // import the gql template function
 const { gql } = require("apollo-server-express");
 const typeDefs = gql`
-type Contractor {
-    _id:ID
-    name:String
-    email:String
-    typeOfServices:String
-    serviceZipcode:String
-    cert:String
-    bio:String
-    photo:String
-    phoneNumber:String
-}
 type User {
     _id:ID
     userName:String
@@ -22,6 +11,8 @@ type User {
     phoneNumber:String
     address:String
     role:String
+    typeOfServices:String
+    serviceZipcode:String
 }
 type Projects {
     _id:ID
@@ -38,25 +29,19 @@ type Offer{
     createdAt:String
 }
 type Query {
-loggedContractor : Contractor
 loggedUser : User
 users:[User]
 user(name:String!): User
-contractors:[Contractor]
-contractor(name:String!): Contractor
 projects(yourName: String): [Projects]
 project(_id: ID!): Projects
 }
 type Mutation {
     login (email:String!, password: String!): Auth
-    ulogin (email:String!, password: String!): Auth
-    addContractor(name:String!,email:String!, password:String!,typeOfServices:String!,address:String!,driverLicense:String!,cert:String,serviceZipcode:String!, bio:String,photo:String,phoneNumber:String!): Auth
-    addUser(userName:String!, firstName:String!, lastName:String!, email:String!, password:String!, phoneNumber:String!, address:String!,role:String): Auth
+    addUser(userName:String!, firstName:String!, lastName:String!, email:String!, password:String!, phoneNumber:String!, address:String!,role:String,typeOfServices:String,serviceZipcode:String ): Auth
     addOffer(projectId: ID!,newOffer:String! ):Projects
 }
 type Auth {
     token: ID!
-    contractor: Contractor
     user: User
 }
 `;
